修复文件上传bug
1.更换图片之后，不能显示原图片
2.上传长款为double时，图片上传失败

+++ src/com/bxb/modules/global/controller/AttachmentUploadController.java
@@ -242,8 +242,8 @@ public class AttachmentUploadController extends BaseController {
 		MultipartHttpServletRequest multipartRequest = (MultipartHttpServletRequest) request;
 
 		ThumbParam tp = new ThumbParam();
-		tp.setWidth(Integer.parseInt(w));
-		tp.setHeight(Integer.parseInt(h));
+		tp.setWidth((int) Double.parseDouble(w));
+		tp.setHeight((int) Double.parseDouble(h));
 
 		boolean isCompress = false;
